# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Roleplay Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /personas:
    get:
      operationId: listPersonas
      tags: [personas]
      summary: List my personas
      responses:
        "200":
          description: Personas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Persona"
    post:
      operationId: createPersona
      tags: [personas]
      summary: Create a persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
  /personas/search:
    get:
      operationId: searchPersonas
      tags: [personas]
      summary: Search all personas
      parameters:
        - name: q
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Personas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Persona"
  /personas/{id}:
    get:
      operationId: getPersona
      tags: [personas]
      summary: Get a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Persona
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
    patch:
      operationId: updatePersona
      tags: [personas]
      summary: Update a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
    delete:
      operationId: deletePersona
      tags: [personas]
      summary: Delete a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /personas/{id}/main:
    post:
      operationId: setMainPersona
      tags: [personas]
      summary: Set persona as main
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"

  # ─── Tags ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Persona:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: string
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        lore:
          type: string
        tags:
          type: array
          items:
            type: string
        isMain:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, ownerUserId, name, age, personality, description, lore, tags, isMain, createdAt]
    CreatePersonaBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        lore:
          type: string
        tags:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
      required: [name, age, personality, description, lore]
    UpdatePersonaBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        personality:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        lore:
          type: ["string", "null"]
        tags:
          type: ["array", "null"]
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
...